refactor(client-3d): extract DreamAudioPlayer helpers to audio/dreamAudio/
Phase 3F: split DreamAudioPlayer.ts 1113→779 LOC by moving pure
helpers into a sibling directory:
- dreamAudio/buffers.ts — generateIR / generateBrightIR /
  generateNoiseBuffer (AudioContext-parameterized pure functions).
- dreamAudio/beatDetection.ts — BeatDetector class encapsulating kick
  + snare onset state, BPM/confidence estimation, and a resetOnsetHistory
  hook for track changes. Sidechain ducking remains in the player via
  an onKick callback.
- dreamAudio/trackLoader.ts — loadAudioTrack + seekToRandomPosition as
  standalone async functions.

Player class keeps effects-chain graph construction, playback
scheduling, and lifecycle — the pieces that genuinely need its
instance state.
